crypto-trader/requirements.txt
flask
flask-cors
PyJWT
requests
python-dotenv
